Relocate common modules
