+++ src/common/constants/color.js
