+++ src/common/constants/icon.js
