--- src/constants/index.js
@@ -1,7 +1,0 @@
-import COLOR from "@/constants/color";
-import ICON from "@/constants/icon";
-
-export default {
-  ...COLOR,
-  ...ICON
-};
